+++ .devcontainer/docker-compose.yml
@@ -1,7 +1,7 @@
 version: "3"
 
 services:
-  pyquail:
+  hexkit:
     build:
       context: .
       dockerfile: ./Dockerfile
