+++ setup.cfg
@@ -14,8 +14,8 @@
 # limitations under the License.
 
 [metadata]
-name = pyquail
-version = attr: pyquail.__version__
+name = hexkit
+version = attr: hexkit.__version__
 description = A library that contains the basic chassis functionality used in services of GHGA
 url = https://github.com/ghga-de/ghga-service-chassis-lib
 long_description = file: README.md
